refactor(runtime): finish inter-agent tool extraction into agent module
Move agent_send, agent_spawn, agent_list, agent_kill plus their two
private helpers (build_agent_manifest_toml, tools_to_parent_capabilities)
into tool_runner/agent.rs, joining the agent_find tool that was
extracted earlier. The module docstring is updated to reflect that
the whole inter-agent surface now lives there.

AGENT_CALL_DEPTH (task-local) is promoted from `static` to
`pub(super) static` so tool_agent_send can keep using its scoped
recursion guard from outside mod.rs. The companion `pub fn
current_agent_depth()` accessor in mod.rs is untouched, so any
external caller continues to resolve `tool_runner::current_agent_depth`.

Taint check on outgoing messages and self-send guard preserved in
order. The stale doc-comment for notify_owner that was orphaned in
mod.rs after the notify extraction is removed in the same hunk.

mod.rs: 10,699 -> 10,461 LOC (-238). cargo check + cargo clippy --lib
both clean.

Refs #3710
